+++ src/main/java/graduation/spendiary/domain/bank/OpenBankTokenRequest.java
@@ -1,11 +1,11 @@
-package graduation.spendiary.security.openbank;
+package graduation.spendiary.domain.bank;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
-public class BankRequestToken {
+public class OpenBankTokenRequest {
     private String code;
     private String client_id;
     private String client_secret;
